Tests: Modified test infrastructure
- Instead of launching a server daemon, simply runs the command line
- Sets TASKDDATA ahead of each run
