+++ test/basetest/__init__.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from .task import Task
 from .taskd import Taskd
 from .testing import TestCase, ServerTestCase
